Exercise 4: Added gorm scopes

+++ 4/models/cart.go
@@ -4,6 +4,6 @@ import "gorm.io/gorm"
 
 type Cart struct {
 	gorm.Model
-	TotalCost float32    `json:"total_cost"`
+	TotalCost float64    `json:"total_cost"`
 	Products  []*Product `gorm:"many2many:cart_products;" json:"products"`
 }
