+++ 4/models/product.go
@@ -8,7 +8,7 @@ type Product struct {
 	gorm.Model
 	Name        string  `json:"name"`
 	Description string  `json:"description"`
-	CategoryID  uint    `json:"category_id"`
-	Price       float32 `json:"price"`
-	Stock       uint    `json:"stock"`
+	CategoryID  uint64  `json:"category_id"`
+	Price       float64 `json:"price"`
+	Stock       uint64  `json:"stock"`
 }
